backend/requirements.txt
# Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9

# Auth - NOME CORRETO AQUI!
djangorestframework-simplejwt==5.3.0

# Database & Search
django-filter==23.3

# Media
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0

# WebSocket
channels==4.0.0
channels-redis==4.2.0

# API Docs
drf-yasg==1.21.7

# Utils
celery==5.3.4
redis==5.0.1
django-ratelimit==4.1.0
drf-nested-routers==0.93.5

# Tests
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
model-bakery==1.17.1
django-extensions